refactor: extract cosine similarity utility
