+++ lib/classifySector.js
@@ -7,20 +7,7 @@ const CLAIRFIELD_SECTORS = [
   'Software, tech & digital'
 ];
 
-function cosineSimilarity(a, b) {
-  let dot = 0;
-  let normA = 0;
-  let normB = 0;
-  for (let i = 0; i < a.length; i++) {
-    const x = a[i];
-    const y = b[i];
-    dot += x * y;
-    normA += x * x;
-    normB += y * y;
-  }
-  if (!normA || !normB) return 0;
-  return dot / (Math.sqrt(normA) * Math.sqrt(normB));
-}
+const cosineSimilarity = require('../utils/cosineSimilarity');
 
 function classifySectorFromVectors(articleVec, sectorVectors, names = CLAIRFIELD_SECTORS) {
   if (!articleVec || !Array.isArray(sectorVectors)) return '';
