+++ routes/articles.js
@@ -5,20 +5,7 @@ const configDb = require('../configDb');
 const { getCompleted } = require('../lib/enrichment/steps');
 const createPipeline = require('../lib/enrichment/pipeline');
 
-function cosineSimilarity(a, b) {
-  let dot = 0;
-  let normA = 0;
-  let normB = 0;
-  for (let i = 0; i < a.length; i++) {
-    const x = a[i];
-    const y = b[i];
-    dot += x * y;
-    normA += x * x;
-    normB += y * y;
-  }
-  if (!normA || !normB) return 0;
-  return dot / (Math.sqrt(normA) * Math.sqrt(normB));
-}
+const cosineSimilarity = require('../utils/cosineSimilarity');
 
 const router = express.Router();
 const openai = new OpenAI({ apiKey: process.env.OPENAI_API_KEY });
